d/control-in: Make Ubuntu qemu-system-common depend on qemu-block-extra
Same reasoning as with qemu-img, there are many common cases which
just should work out of the box.

Note: this was discussed with Debian back then but they were not
interested. Our reasoning was very strong around the Ubuntu OpenStack
usage which isn't that important for Debian. Therefore this change is
Ubuntu-only.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>