d/control*, d/rules: disable xen by default, but provide universe package qemu-system-x86-xen as alternative
xen went to universe in Ubuntu and in general we wanted to split
support/bug handling of the much less exercised xen from kvm use cases.

Therefore in Ubuntu the qemu-system-x86 in main can not depend on xen
anymore, but we provide a separate qemu-system-x86-xen which fulfills
the xen-hipervisor needs and at the same time keeps the more commonly
used qemu-system-x86 free of xen code.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>