save block modules on upgrades (LP: #1847361)
After upgrade a still running qemu of a former version can't load the
new modules e.g. for extended storage support. Qemu 5.0 has the code to
allow defining a path that it will load these modules from.

Augment the postrm/prerm code to save the current drivers to be able to
fall back to these. The behavior will be:
- on upgrades save modules to /run
- on removal/purge clean the /run (user wants to get rid of qemu)
- on reboot /run is auto-cleared which is great since all old qemu
  processes are gone as well (much better than a complex logic)
- for now we don't care about running in a chroot/debootstrap
  environment, we save modules unconditionally in /run

Original idea and implementation is by Christian Ehrhardt