nbd-server-avoid-long-error-message-assertions-CVE-2020-10761.patch
CVE-2020-10761: An assertion failure issue was found in the NBD Server
in all QEMU versions before QEMU 5.0.1. This flaw occurs when an nbd-client
sends a spec-compliant request that is near the boundary of maximum permitted
request length. A remote nbd-client could use this flaw to crash the
qemu-nbd server resulting in a denial of service.