qemu-user: attempt to preserve argv[0] when run under binfmt (#970460, #959530)
This is an attempt to fix argv[0] trashing when a foreign
executable is run using in-kernel binfmt-misc mechanism.

See comments in linux-user/main.c (in the patch) for details.

Please note this change also changes the enterpreter name
for all linux-user registered binfmts, so it potentially
can break stuff.

The actual binary name registered with binfmt-misc is
/usr/libexec/qemu-binfmt/foo-binfmt-P, which is a symlink
to /usr/lib/qemu-foo[-static].

Let's see..