enable /run/qemu mount on ubuntu only
qemu-system can load (block for now) modules after startup
by means of device hotplug mechanism. When qemu and modules
are upgraded, but running instances are not, running instances
(which are older qemu) can't load new modules from disk.
For this to work, during upgrade we copy old modules to
/run/qemu/$version, and running (old) qemu can load modules
from there.  But in order to be able to do that, this directory
should be mounted without -o noexec. By default on debian whole
/run is mounted -o noexec, so this does not work by default.
qemu-block-extra provides a systemd mount unit run-qemu.mount
which mounts /run/qemu as a separate tmpfs without -o noexec.
On debian we do not enable this mount unit by default since
apparently this is against policy. But ubuntu wants to be
more user-friendly and enables it by default. Do this by
running dh_installsystemd with arguments depending on $VENDOR.